refactor: settings module.
