+++ packages/ui/src/home.tsx
@@ -8,10 +8,10 @@ import styled from 'styled-components'
 
 import { useGlobals } from './globals.js'
 import { useVehiclesDispatch } from './contexts/vehicles.js'
-import { useTheme } from './contexts/settings/theme.js'
+import { useTheme } from './modules/settings/contexts/theme.js'
 import { BusSelector } from './components/busSelector.js'
 import { Loading } from './components/loading.js'
-import { Settings } from './components/settings/index.js'
+import { Settings } from './modules/settings/components/settings.js'
 import { Favorites } from './modules/favorites/components/favorites.js'
 import { Info } from './components/info.js'
 import { Predictions } from './components/predictions.js'
