+++ packages/ui/src/modules/settings/components/theme.tsx
@@ -1,10 +1,10 @@
 import styled from 'styled-components'
 import { useCallback } from 'react'
 
-import { FormItem } from '../formItem.js'
-import { useStorageDispatch } from '../../contexts/storage.js'
-import { useTheme } from '../../contexts/settings/theme.js'
-import { isAMode } from '../../contexts/util.js'
+import { FormItem } from '../../../components/formItem.js'
+import { useStorageDispatch } from '../../../contexts/storage.js'
+import { useTheme } from '../contexts/theme.js'
+import { isAMode } from '../util.js'
 
 import type { FC, ChangeEvent } from 'react'
