+++ packages/ui/src/modules/settings/types.ts
@@ -1,0 +1,5 @@
+type Mode = 'light' | 'dark'
+type SpeedUnit = 'kph' | 'mph'
+type PredictionFormat = 'time' | 'minutes'
+
+export type { Mode, SpeedUnit, PredictionFormat }
